Celebrity-classification/requirements.txt
catboost==1.0.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
dlib==19.23.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.3
Flask-WTF==1.0.0
fonttools==4.30.0
graphviz==0.19.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
kiwisolver==1.4.0
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
plotly==5.6.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
scipy==1.8.0
six==1.16.0
tenacity==8.0.1
urllib3==1.26.9
Werkzeug==2.0.3
WTForms==3.0.1
